+++ fuzzers/fuzzbench_forkserver_cmplog/Cargo.toml
@@ -22,5 +22,6 @@ which = "4.4"
 [dependencies]
 libafl = { path = "../../libafl/" }
 libafl_bolts = { path = "../../libafl_bolts/" }
+libafl_targets = { path = "../../libafl_targets/" }
 clap = { version = "4.0", features = ["default"] }
 nix = "0.26"
