+++ libafl_targets/src/cmplog.h
@@ -57,28 +57,4 @@ void __libafl_targets_cmplog_routines(uintptr_t k, const uint8_t *ptr1,
 void __libafl_targets_cmplog_routines_len(uintptr_t k, const uint8_t *ptr1,
                                           const uint8_t *ptr2, size_t len);
 
-static inline void __libafl_targets_cmplog(uintptr_t k, uint8_t shape,
-                                           uint64_t arg1, uint64_t arg2) {
-  if (!libafl_cmplog_enabled) { return; }
-  libafl_cmplog_enabled = false;
-
-  uint16_t hits;
-  if (libafl_cmplog_map_ptr->headers[k].kind != CMPLOG_KIND_INS) {
-    libafl_cmplog_map_ptr->headers[k].kind = CMPLOG_KIND_INS;
-    libafl_cmplog_map_ptr->headers[k].hits = 1;
-    libafl_cmplog_map_ptr->headers[k].shape = shape;
-    hits = 0;
-  } else {
-    hits = libafl_cmplog_map_ptr->headers[k].hits++;
-    if (libafl_cmplog_map_ptr->headers[k].shape < shape) {
-      libafl_cmplog_map_ptr->headers[k].shape = shape;
-    }
-  }
-
-  hits &= CMPLOG_MAP_H - 1;
-  libafl_cmplog_map_ptr->vals.operands[k][hits].v0 = arg1;
-  libafl_cmplog_map_ptr->vals.operands[k][hits].v1 = arg2;
-  libafl_cmplog_enabled = true;
-}
-
 #endif
