+++ libafl_targets/src/cmps/observers/mod.rs
@@ -1,0 +1,6 @@
+pub mod cmplog;
+pub use cmplog::*;
+
+/// aflpp style cmplog observer
+pub mod aflpp;
+pub use aflpp::*;
