+++ libafl_targets/src/cmps/stages/mod.rs
@@ -1,0 +1,3 @@
+/// cmplog tracing for aflpp style cmplog
+pub mod aflpptracing;
+pub use aflpptracing::*;
